backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "passlib>=1.7.4",
    "pgvector>=0.4.1",
    "psycopg[binary,pool]>=3.2.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pymupdf>=1.26.4",
    "python-docx>=1.2.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.43",
    "tiktoken>=0.11.0",
    "uvicorn>=0.35.0",
    "voyageai>=0.3.5",
]

